Tidy up product test

+++ tests/_fixtures/product_fixtures.py
@@ -11,6 +11,7 @@ def product_simple(request_context: APIRequestContext):
 
     yield product_simple
 
-    request_context.delete(
-        f"wp-json/wc/v3/products/{product_simple['id']}", params={"force": True}
-    )
+    if product_simple["id"]:
+        request_context.delete(
+            f"wp-json/wc/v3/products/{product_simple['id']}", params={"force": True}
+        )
